pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "github-org-mirror"
version = "0.1.0"
description = "Two-way sync between local folders and GitHub organizations"
readme = "README.md"
license = {text = "MIT"}
authors = [
    {name = "President Anderson"}
]
requires-python = ">=3.8"
classifiers = [
    "Development Status :: 4 - Beta",
    "Environment :: Console",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.8",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Software Development :: Version Control :: Git",
]
keywords = ["github", "organization", "sync", "mirror", "git"]
dependencies = [
    "click>=8.0.0",
    "watchdog>=3.0.0",
    "pyyaml>=6.0",
    "rich>=13.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "pytest-cov>=4.0.0",
    "black>=23.0.0",
    "ruff>=0.1.0",
]

[project.scripts]
github-org-mirror = "github_org_mirror.cli:main"

[project.urls]
Homepage = "https://github.com/Sovereign-Operating-System/github-org-mirror"
Repository = "https://github.com/Sovereign-Operating-System/github-org-mirror"
Issues = "https://github.com/Sovereign-Operating-System/github-org-mirror/issues"

[tool.setuptools.packages.find]
where = ["."]

[tool.black]
line-length = 100
target-version = ["py38"]

[tool.ruff]
line-length = 100
target-version = "py38"
